+++ templates/lesson.js
@@ -1,0 +1,16 @@
+const getLessonContent = ({ title, url, prev, next }) =>
+  `
+# ${title}
+
+**[📹 Video](${url})**
+
+TODO
+
+${(prev || next) && `---`}
+
+${prev ? `📹 [Go to Previous Lesson](${prev})` : ""}
+${next ? `📹 [Go to Next Lesson](${next})` : ""}
+
+`.trim() + "\n";
+
+module.exports = getLessonContent;
