+++ utils/url.js
@@ -1,0 +1,6 @@
+const baseUrl = "https://egghead.io";
+const baseLessonUrl = `${baseUrl}/lessons/`;
+const baseCourseUrl = `${baseUrl}/courses/`;
+const getCourseSlug = () => courseUrl.replace(baseCourseUrl, "");
+
+module.exports = { baseUrl, baseLessonUrl, baseCourseUrl, getCourseSlug };
